Rearrange memory segment implementation hierarchy
