+++ src/jdk.incubator.foreign/share/classes/jdk/internal/foreign/HeapMemorySegment.java
@@ -1,0 +1,121 @@
+/*
+ *  Copyright (c) 2019, Oracle and/or its affiliates. All rights reserved.
+ *  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ *  This code is free software; you can redistribute it and/or modify it
+ *  under the terms of the GNU General Public License version 2 only, as
+ *  published by the Free Software Foundation.  Oracle designates this
+ *  particular file as subject to the "Classpath" exception as provided
+ *  by Oracle in the LICENSE file that accompanied this code.
+ *
+ *  This code is distributed in the hope that it will be useful, but WITHOUT
+ *  ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ *  FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ *  version 2 for more details (a copy is included in the LICENSE file that
+ *  accompanied this code).
+ *
+ *  You should have received a copy of the GNU General Public License version
+ *  2 along with this work; if not, write to the Free Software Foundation,
+ *  Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ *   Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ *  or visit www.oracle.com if you need additional information or have any
+ *  questions.
+ *
+ */
+
+package jdk.internal.foreign;
+
+import jdk.incubator.foreign.MemorySegment;
+import jdk.internal.access.JavaNioAccess;
+import jdk.internal.access.SharedSecrets;
+import jdk.internal.misc.Unsafe;
+import jdk.internal.vm.annotation.ForceInline;
+
+import java.nio.ByteBuffer;
+import java.util.Objects;
+
+/**
+ * This class provides an immutable implementation for the {@code MemorySegment} interface. This class contains information
+ * about the segment's spatial and temporal bounds, as well as the addressing coordinates (base + offset) which allows
+ * unsafe access; each memory segment implementation is associated with an owner thread which is set at creation time.
+ * Access to certain sensitive operations on the memory segment will fail with {@code IllegalStateException} if the
+ * segment is either in an invalid state (e.g. it has already been closed) or if access occurs from a thread other
+ * than the owner thread. See {@link MemoryScope} for more details on management of temporal bounds.
+ */
+public final class HeapMemorySegment extends AbstractMemorySegment {
+
+    private static final Unsafe UNSAFE = Unsafe.getUnsafe();
+    private static final int BYTE_ARR_BASE = UNSAFE.arrayBaseOffset(byte[].class);
+
+    final long offset;
+    final Object base;
+
+    @ForceInline
+    HeapMemorySegment(long offset, Object base, long length, int mask, Thread owner, MemoryScope scope) {
+        super(length, mask, owner, scope);
+        this.offset = offset;
+        this.base = Objects.requireNonNull(base);
+    }
+
+    @Override
+    AbstractMemorySegment dup(long offset, long size, int mask, Thread owner, MemoryScope scope) {
+        return new HeapMemorySegment(this.offset + offset, base, size, mask, owner, scope);
+    }
+
+    @Override
+    ByteBuffer makeByteBuffer() {
+        if (!(base() instanceof byte[])) {
+            throw new UnsupportedOperationException("Not an address to an heap-allocated byte array");
+        }
+        JavaNioAccess nioAccess = SharedSecrets.getJavaNioAccess();
+        return nioAccess.newHeapByteBuffer((byte[]) base(), (int)min() - BYTE_ARR_BASE, (int) byteSize(), this);
+    }
+
+    @Override
+    long min() {
+        return offset;
+    }
+
+    @Override
+    Object base() {
+        return base;
+    }
+
+    // factories
+
+    public static MemorySegment makeArraySegment(byte[] arr) {
+        return makeArraySegment(arr, arr.length, Unsafe.ARRAY_BYTE_BASE_OFFSET, Unsafe.ARRAY_BYTE_INDEX_SCALE);
+    }
+
+    public static MemorySegment makeArraySegment(char[] arr) {
+        return makeArraySegment(arr, arr.length, Unsafe.ARRAY_CHAR_BASE_OFFSET, Unsafe.ARRAY_CHAR_INDEX_SCALE);
+    }
+
+    public static MemorySegment makeArraySegment(short[] arr) {
+        return makeArraySegment(arr, arr.length, Unsafe.ARRAY_SHORT_BASE_OFFSET, Unsafe.ARRAY_SHORT_INDEX_SCALE);
+    }
+
+    public static MemorySegment makeArraySegment(int[] arr) {
+        return makeArraySegment(arr, arr.length, Unsafe.ARRAY_INT_BASE_OFFSET, Unsafe.ARRAY_INT_INDEX_SCALE);
+    }
+
+    public static MemorySegment makeArraySegment(float[] arr) {
+        return makeArraySegment(arr, arr.length, Unsafe.ARRAY_FLOAT_BASE_OFFSET, Unsafe.ARRAY_FLOAT_INDEX_SCALE);
+    }
+
+    public static MemorySegment makeArraySegment(long[] arr) {
+        return makeArraySegment(arr, arr.length, Unsafe.ARRAY_LONG_BASE_OFFSET, Unsafe.ARRAY_LONG_INDEX_SCALE);
+    }
+
+    public static MemorySegment makeArraySegment(double[] arr) {
+        return makeArraySegment(arr, arr.length, Unsafe.ARRAY_DOUBLE_BASE_OFFSET, Unsafe.ARRAY_DOUBLE_INDEX_SCALE);
+    }
+
+    private static MemorySegment makeArraySegment(Object arr, int size, int base, int scale) {
+        MemoryScope scope = new MemoryScope(null, null);
+        long length = size * scale;
+        return new HeapMemorySegment(base, arr, length,
+                defaultAccessModes(length), Thread.currentThread(), scope);
+    }
+}
