+++ src/jdk.incubator.foreign/share/classes/jdk/internal/foreign/MappedMemorySegment.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright (c) 2020, Oracle and/or its affiliates. All rights reserved.
+ * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
+ *
+ * This code is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU General Public License version 2 only, as
+ * published by the Free Software Foundation.  Oracle designates this
+ * particular file as subject to the "Classpath" exception as provided
+ * by Oracle in the LICENSE file that accompanied this code.
+ *
+ * This code is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
+ * FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License
+ * version 2 for more details (a copy is included in the LICENSE file that
+ * accompanied this code).
+ *
+ * You should have received a copy of the GNU General Public License version
+ * 2 along with this work; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
+ *
+ * Please contact Oracle, 500 Oracle Parkway, Redwood Shores, CA 94065 USA
+ * or visit www.oracle.com if you need additional information or have any
+ * questions.
+ */
+
+package jdk.internal.foreign;
+
+import jdk.incubator.foreign.MemorySegment;
+import jdk.internal.access.foreign.UnmapperProxy;
+import sun.nio.ch.FileChannelImpl;
+
+import java.io.IOException;
+import java.nio.channels.FileChannel;
+import java.nio.file.OpenOption;
+import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
+
+public class MappedMemorySegment extends NativeMemorySegment {
+
+    MappedMemorySegment(long min, long length, int mask, Thread owner, MemoryScope scope) {
+        super(min, length, mask, owner, scope);
+    }
+
+    // create and map a file into a fresh segment
+    public static MemorySegment makeMappedSegment(Path path, long bytesSize, FileChannel.MapMode mapMode) throws IOException {
+        if (bytesSize <= 0) throw new IllegalArgumentException("Requested bytes size must be > 0.");
+        try (FileChannelImpl channelImpl = (FileChannelImpl)FileChannel.open(path, openOptions(mapMode))) {
+            UnmapperProxy unmapperProxy = channelImpl.mapInternal(mapMode, 0L, bytesSize);
+            MemoryScope scope = new MemoryScope(null, unmapperProxy::unmap);
+            return new MappedMemorySegment(unmapperProxy.address(), defaultAccessModes(bytesSize),
+                    AbstractMemorySegment.DEFAULT_MASK, Thread.currentThread(), scope);
+        }
+    }
+
+    private static OpenOption[] openOptions(FileChannel.MapMode mapMode) {
+        if (mapMode == FileChannel.MapMode.READ_ONLY) {
+            return new OpenOption[] { StandardOpenOption.READ };
+        } else if (mapMode == FileChannel.MapMode.READ_WRITE || mapMode == FileChannel.MapMode.PRIVATE) {
+            return new OpenOption[] { StandardOpenOption.READ, StandardOpenOption.WRITE };
+        } else {
+            throw new UnsupportedOperationException("Unsupported map mode: " + mapMode);
+        }
+    }
+}
