+++ src/jdk.incubator.foreign/share/classes/jdk/internal/foreign/MemoryScope.java
@@ -33,7 +33,7 @@ import java.lang.invoke.VarHandle;
 
 /**
  * This class manages the temporal bounds associated with a memory segment. A scope has a liveness bit, which is updated
- * when the scope is closed (this operation is triggered by {@link MemorySegmentImpl#close()}). Furthermore, a scope is
+ * when the scope is closed (this operation is triggered by {@link NativeMemorySegment#close()}). Furthermore, a scope is
  * associated with an <em>atomic</em> counter which can be incremented (upon calling the {@link #acquire()} method),
  * and is decremented (when a previously acquired segment is later closed).
  */
